# [DRAFT] Reduce number of call to makeVarBinds/resolveWithMib