--- a/snmp/datadog_checks/snmp/models.py
+++ b/snmp/datadog_checks/snmp/models.py
@@ -24,8 +24,6 @@ class OID(object):
     def __init__(self, value):
         # type: (Union[Sequence[int], str, ObjectName, ObjectIdentity, ObjectType]) -> None
         parts = None  # type: Optional[Tuple[int, ...]]
-        self._value = value
-        print("value type", type(self._value))
         try:
             parts = parse_as_oid_tuple(value)
         except CouldNotDecodeOID: