### What does this PR do?

Reduce number of call to makeVarBinds/resolveWithMib

### Motivation

- Simplify logic using tuple oids
- Possible small performance improvement (some variations between bench runs)

Before:
```
§ for i in `seq 1 5`; do ddev test --bench snmp --skip-env -k test_profile_f5 | grep -A 4 -- '--- benchmark'; done
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.2187  1.4112  1.2778  0.0805  1.2337  0.0986       1;0  0.7826       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1714  1.3520  1.2314  0.0696  1.2112  0.0476       1;1  0.8121       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1920  1.3250  1.2381  0.0532  1.2318  0.0668       1;0  0.8077       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1725  1.2341  1.2023  0.0219  1.2003  0.0187       2;0  0.8317       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.2842  1.4184  1.3331  0.0538  1.3326  0.0711       1;0  0.7501       5           1
---------------------------------------------------------------------------------------------------------
```

After (With this PR):
```
§ for i in `seq 1 5`; do ddev test --bench snmp --skip-env -k test_profile_f5 | grep -A 4 -- '--- benchmark'; done
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1439  1.2869  1.2048  0.0543  1.2098  0.0690       2;0  0.8300       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1459  1.2089  1.1676  0.0268  1.1554  0.0397       1;0  0.8565       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1369  1.2050  1.1724  0.0267  1.1800  0.0399       2;0  0.8530       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1550  1.2383  1.1887  0.0392  1.1700  0.0705       1;0  0.8413       5           1
---------------------------------------------------------------------------------------------------------
------------------------------------------- benchmark: 1 tests ------------------------------------------
Name (time in s)        Min     Max    Mean  StdDev  Median     IQR  Outliers     OPS  Rounds  Iterations
---------------------------------------------------------------------------------------------------------
test_profile_f5      1.1498  1.2109  1.1693  0.0257  1.1549  0.0331       1;0  0.8552       5           1
---------------------------------------------------------------------------------------------------------
```